Convert Numpy Types

+++ skvideo/measure/scene.py
@@ -21,7 +21,7 @@ def _percentage_distance(canny_in, canny_out, r):
     E_1 = scipy.ndimage.morphology.binary_dilation(canny_in, structure=diamond, iterations=r)
     E_2 = scipy.ndimage.morphology.binary_dilation(canny_out, structure=diamond, iterations=r)
 
-    return 1.0 - np.float(np.sum(E_1 & E_2))/np.float(np.sum(E_1))
+    return 1.0 - float(np.sum(E_1 & E_2))/float(np.sum(E_1))
 
 def _scenedet_edges(videodata, threshold):
     # the first frame is always a new scene
@@ -72,8 +72,8 @@ def _scenedet_intensity(videodata, parameter1):
         hist1, bins = np.histogram(luminancedata[t], bins=256, range=(0, 255))
         hist2, bins = np.histogram(luminancedata[t+1], bins=256, range=(0, 255))
 
-        hist1 = hist1.astype(np.float)
-        hist2 = hist2.astype(np.float)
+        hist1 = hist1.astype(float)
+        hist2 = hist2.astype(float)
 
         hist1 /= 256.0
         hist2 /= 256.0
